split large handler.go file into smaller files
