+++ apps/bank/transactions_handler.go
@@ -1,0 +1,110 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"strconv"
+	"strings"
+
+	acpClient "github.com/cloudentity/acp-client-go/models"
+	"github.com/cloudentity/openbanking-quickstart/models"
+	"github.com/gin-gonic/gin"
+	"github.com/go-openapi/strfmt"
+)
+
+// swagger:parameters getTransactionsRequest
+type GetTransactionsRequest struct {
+	RequestHeaders
+}
+
+// swagger:route GET /transactions bank getTransactionsRequest
+//
+// get transactions
+//
+// Security:
+//   defaultcc: accounts
+//
+// Responses:
+//   200: OBReadTransaction6
+//   400: OBErrorResponse1
+//   403: OBErrorResponse1
+//   404: OBErrorResponse1
+func (s *Server) GetTransactions() func(ctx *gin.Context) {
+	return func(c *gin.Context) {
+		var (
+			introspectionResponse *acpClient.IntrospectOpenbankingAccountAccessConsentResponse
+			userTransactions      []models.OBTransaction6
+			err                   error
+		)
+
+		if introspectionResponse, err = s.IntrospectAccountsToken(c); err != nil {
+			msg := fmt.Sprintf("failed to introspect token: %+v", err)
+			c.JSON(http.StatusBadRequest, models.OBErrorResponse1{
+				Message: &msg,
+			})
+			return
+		}
+
+		grantedPermissions := introspectionResponse.Permissions
+
+		scopes := strings.Split(introspectionResponse.Scope, " ")
+		if !has(scopes, "accounts") {
+			msg := "token has no accounts scope granted"
+			c.JSON(http.StatusForbidden, models.OBErrorResponse1{
+				Message: &msg,
+			})
+			return
+		}
+
+		if !has(grantedPermissions, "ReadTransactionsBasic") {
+			msg := "ReadTransactionsBasic permission has not been granted"
+			c.JSON(http.StatusForbidden, models.OBErrorResponse1{
+				Message: &msg,
+			})
+			return
+		}
+
+		if userTransactions, err = s.Storage.GetTransactions(introspectionResponse.Sub); err != nil {
+			msg := err.Error()
+			c.JSON(http.StatusNotFound, models.OBErrorResponse1{
+				Message: &msg,
+			})
+			return
+		}
+
+		transactions := []*models.OBTransaction6{}
+
+		for _, transaction := range userTransactions {
+			t := transaction
+			if has(introspectionResponse.AccountIDs, string(*t.AccountID)) {
+				if !has(grantedPermissions, "ReadTransactionsDetail") {
+					t.TransactionInformation = ""
+					t.Balance = &models.OBTransactionCashBalance{}
+					t.MerchantDetails = &models.OBMerchantDetails1{}
+					t.CreditorAgent = &models.OBBranchAndFinancialInstitutionIdentification61{}
+					t.CreditorAccount = &models.OBCashAccount60{}
+					t.DebtorAgent = &models.OBBranchAndFinancialInstitutionIdentification62{}
+					t.DebtorAccount = &models.OBCashAccount61{}
+				}
+
+				transactions = append(transactions, &t)
+			}
+		}
+
+		self := strfmt.URI(fmt.Sprintf("http://localhost:%s/transactions", strconv.Itoa(s.Config.Port)))
+
+		response := models.OBReadTransaction6{
+			Data: &models.OBReadDataTransaction6{
+				Transaction: transactions,
+			},
+			Meta: &models.Meta{
+				TotalPages: int32(len(transactions)),
+			},
+			Links: &models.Links{
+				Self: &self,
+			},
+		}
+
+		c.PureJSON(http.StatusOK, response)
+	}
+}
